requirements.txt
requests
pandas
matplotlib
seaborn
scikit-learn
streamlit
jupyter
python-dotenv
xlsxwriter
